fix: a parked run is readable like every other (p0445)
Operator spotted it: the "Needs you" card had no right-hand side — no story
spine, no step position, no cost, no elapsed — while every finished row below it
carried all four. So the single run that actually needs a decision was the only
one that could not be judged without opening it. Live run 9a30 stood at 36/48,
$4.58 and 27 minutes in, and the card said none of it.

The columns move out of RunRow into a shared RunStats rather than being
re-implemented on the card: a second hand-built copy of the same four facts is
the one that drifts. RunRow keeps its finished-pill fallback through a
spineFallback slot, so what it renders is unchanged.

The card keeps its own layout. It exists because a parked run is answerable in
place, and that affordance is the point — what was missing was the facts, not
the row shape.

620 dashboard tests green.
