+++ .agentsmith/contexts/default/context.yaml
@@ -75,6 +75,7 @@ state:
   done:
     p0434: "THE PR STEP SURVIVES A PROVIDER BLIP (backend). Discovered by p0433, not by a run: declaring CommitAndPR a MODEL step made visible that on the path where it reaches the account first, PhaseAccounting TAKES rather than reuses — so the step that opens the pull request makes a provider call at the last step of the run, after every unit of work is done. The p0350 shape, which cost two draft PRs the one time it happened for real. Fix: a failure taking the account becomes an account that COULD NOT BE TAKEN — the same shape TakeAsync already returns for a failed delivery diff — so the gate still refuses the run (p0420) but refuses it WITH a pull request to look at. Rejected: a retry (adds latency to the same failure) and moving the account earlier (only helps the path that was never exposed). Cancellation still propagates, filtered on the token, because swallowing it would turn an operator cancel into a silent unaccounted delivery. FOUND WHILE BUILDING: PhaseAccounting sat at 116 of its 120 ratchet lines so the try/catch did not fit; the first extraction, SpecAccountLog.Write(logger, account), was refused by NoStaticStateRuleTests — a static that needs a collaborator is a service without a constructor. Inverted: the static composes LINES, the caller writes them. -> .agentsmith/phases/done/p0434-the-pr-step-survives-a-provider-blip.yaml"
     p0432: "THE SUITE MIGRATES ONCE (tests). 41 sites called Database.Migrate(); the cost was O(classes) and it had already cost two CI cycles whose victims were NOT database tests — on a two-core runner the migration burst starves async continuations, and the casualty is whichever test asserts a wall-clock bound (#501 a 50ms send bound, #502 a monotonic-gap window). The stores are `Data Source=:memory:`, so the spec's file-copy did not apply — SQLite RESTORES a database from another connection far faster than it builds one, which is the same O(1) win by a different mechanism (MigratedStoreTemplate, lazy template + BackupDatabase behind a lock, since connections are not thread-safe and xUnit runs classes in parallel). A shared store was rejected outright: a copy is not a share, and isolation is what the 39 separate stores bought. Adopted by rewriting ONLY the exact two-line shape (open private in-memory store, then migrate) across 21 classes; the rest build stores differently and are worth converting when someone is already in the file. The rewrite hit the template's OWN migration method on the first pass — caught at once, and the reason to keep it narrow. Third step done in the same phase: both widened hang-detectors back from 120s to 30s, because leaving them would keep the stopgap's cost (a real deadlock taking two minutes to surface) after removing the pressure that justified it. 3866 green. -> .agentsmith/phases/done/p0432-the-suite-migrates-once.yaml"
+    p0445: "A PARKED RUN IS READABLE LIKE EVERY OTHER (dashboard). OPERATOR SPOTTED IT: the 'Needs you' card had no right-hand side — no story spine, no step position, no cost, no elapsed — while every finished row below it carried all four. So the single run that actually needs a decision was the only one that could not be judged without opening it: live run 9a30 stood at 36/48, $4.58 and 27 minutes in, and the card said none of it. The columns move out of RunRow into a shared RunStats rather than being re-implemented, because a second hand-built copy of the same four facts is the one that drifts; RunRow keeps its finished-pill fallback through a spineFallback slot, so its rendering is unchanged. The card keeps its own layout — it exists because a parked run is answerable in place, and that affordance is the point; what was missing was the facts, not the row shape. Grid gains four auto columns, collapsing to the row's own small-width behaviour. 620 dashboard green. -> .agentsmith/phases/done/p0445-a-parked-run-is-readable-like-every-other.yaml"
     p0444: "A PHASE STARTS WITHOUT THE LAST ONE'S REPAIR (backend). FOUND BY THE AGENT MID-RUN, AND IT REFUSED TO OBEY. Live run 9a30: phase c's FIRST master pass was handed p0438's repair block carrying phase B's outstanding criteria ('this is a REPAIR pass ... close exactly these ... the rest of the phase is already accounted for, adding more is scope you were not given') while the tree still held every MassTransit reference phase c existed to remove. It asked instead of complying. THE QUIETER HALF IS WORSE: PhaseRepairAttempted carried over too, so phase c had already spent the single repair p0438 grants it, on another phase's behalf — the symptom would read as 'the repair loop stopped working'. p0438 wrote both keys and nothing cleared them; SelectPhase swapped the spec and left them. Reset on ENTERING a phase (not at the end of VerifyPhase, which does not run when a phase fails outright); named PhaseRepairScope rather than two inline Remove calls. p0440 proved itself in the same run: the park was recorded waiting_for_input, so the question reached the operator instead of being filed as a finished run. 3873 green. -> .agentsmith/phases/done/p0444-a-phase-starts-without-the-last-ones-repair.yaml"
     p0443: "WAIT FOR WHAT IS ASSERTED (tests). CI red on main after the 0.131.1 release merge — and NOT a regression: the identical code passed 28 minutes earlier on the p0442 build, and the release commit touches CHANGELOG.md, version.txt and the release-please manifest, no code. The gap: WaitForTerminalRowAsync polls Runs for a FinishedAt, the assertion counts TRAIL rows — two writes, one wait, so under load the assertion runs between them. The failure named it exactly (persisted true, status success, trail rows 0). NOT the p0432 family though it looks like it: the budget was already 60s and widening it would have bought quiet rather than correctness — a test that waits for the wrong thing is wrong on a fast runner too, it just gets away with it. Establishing it was not a regression first turned 'what did we break' into 'what did this test always assume'. -> .agentsmith/phases/done/p0443-wait-for-what-is-asserted.yaml"
     p0442: "THE EMBEDDED PIN CATCHES UP (backend). Two pins exist and only one was stale: the deployment pins 4.5.0 in config, the EMBEDDED fallback sat at v4.0.0, five releases behind — nothing broke in production, which is why it went unnoticed. p0415 PREDICTED THE MOMENT: its comment said ownership of spec-derivation-master moves to the catalog 'when a parity release exists AND the pin moves — NoPromptName_HasTwoOwners fails at exactly that moment', and it fired on the first build after the bump. Both reasons p0415 gave had expired: v4.5.0 carries the cut-sizing rule (p0413) and ships_code left the model in p0421. CHECKED BEFORE DELETING: a line diff suggested the catalog had LOST the cut-sizing rule, which would have made the flip a regression — it had not, the rule is reordered, and the catalog's criterion rule is the EVOLVED form with frontmatter the embedded copy never had. The rules are now asserted against the PACKAGED catalog (that is what makes a pin bump a decision rather than a version string) with WHITESPACE COLLAPSED, because 'the branch diff' straddles a wrap in v4.5.0 and a rule that is present must not read as missing. The tar walk is shared rather than copied — same lesson as p0441 an hour earlier. Tarball checksum verified against the published sidecar; the build re-verifies on every compile. 3870 green. -> .agentsmith/phases/done/p0442-the-embedded-pin-catches-up.yaml"
